Support fabric swatch export
There is a new "Swatch" section, intended to be hidden and shown as a dialog.
It is rendering and updating correctly in all scenarios, as far as I can tell.

I also did some major surgery on the controls, working toward a better
organization of the UI.  However, styling at this moment is positively atrocious.