Corrected repeat reshape defects
The problem was that refreshEverything() still requires an augmented
version of beadIndexToColor, with the currentRepeat prepended.   That
really needs to get changed, and probably the logic in the two
circumferenceGot*() functions can be simplified.

In fact, the augmented array should be replaced with an object, and
currentRepeat should become a function that returns the length of
bookIndexToColor (which also needs to be renamed).